+++ src/dfsparks/timer.cpp
@@ -10,7 +10,7 @@ namespace dfsparks {
 int32_t timeMillis() {
 #ifndef ARDUINO
   static auto t0 = std::chrono::steady_clock::now();
-  return std::chrono::duration_cast<std::chrono::milliseconds>(
+  return (int32_t)std::chrono::duration_cast<std::chrono::milliseconds>(
              std::chrono::steady_clock::now() - t0)
       .count();
 #else
